clone: clear the frame pointer in the child process on relevant ports
This just mirrors what is done in the start code for the affected ports, as well
as what is already done for the three x86 ports.

Clearing the frame pointer helps protect FP-based unwinders which have no way of
knowing that the FP register should be considered undefined in the child process
portion of clone(). In practice, we found this change to be necessary when
running the Zig standard library tests under qemu-aarch64_be with musl linked.