configure: prevent compilers from turning a * b + c into fma(a, b, c)
clang at least will turn these expressions into llvm.fmuladd.* intrinsic calls
in the generated LLVM IR. This means we're completely at the mercy of the
backend in regards to whether it gets lowered to an fma() call or simply an fmul
+ fadd. Similarly, optimization passes are allowed to perform transformations on
those intrinsic calls that fmul + fadd would not permit. I imagine the story is
similar for gcc.